Move files_sharing Controllers => Controller
lib/Controller is the default location for controllers. So lets put them
all in there.
